+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 10.1.3
 - extension-dapp 0.44.3
 - react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 2.9.3
 - api, api-contract, types 9.0.1